Added correct dependencies between service controller and storage classes
