+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -1,4 +1,21 @@
 package ru.yandex.practicum.filmorate.storage;
 
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.Collection;
+
 public interface UserStorage {
+    Collection<User> findAll();
+
+    User create(User user);
+
+    User update(User newUser);
+
+    void addFriend(long userId, long friendId);
+
+    void removeFriend(long userId, long friendId);
+
+    Collection<User> getFriendList(long userId);
+
+    Collection<User> getCommonFriends(long userId, long secondUserId);
 }
